+++ public/js/shinayuu-index-bundle.js
@@ -281,7 +281,7 @@ var updatePreviewState = {
   installerPath: '',
   installerOpened: false,
   cached: false,
-  currentVersion: '2.0.15',
+  currentVersion: '2.0.16',
   version: '2.0.0',
   configured: false,
   preview: true,